Api/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
psycopg==3.2.3
psycopg2-binary==2.9.10
python-dotenv>=1.0,<2.0
google-generativeai>=0.8.3
celery>=5.3,<5.4
redis>=5.0,<5.1
django-celery-beat>=2.5,<2.6
django-celery-results>=2.5,<2.6
gunicorn>=21.0,<22.0
whitenoise>=6.5,<7.0
